v017
I created a new PCB boolean to reflect the new board and so the Dorsal
Cover fits tightly around the connectors. Added Cable Port boolean to
PCB tray so that it doesn't prevent a user from fully inserting the
connectors. Slightly lowered the proximal plane of the Master - Dorsal
block. This puts it more inline with the thumb joint and allows the
thumb back stop to be bigger without self-intersecting.