Fix STL cable port glitch
The PCB connector boolean was not applying correctly. Tabbing in and out
of edit mode on the PCB boolean fixed this. I saved out the corrected
STLs for the back cover for a left and right hand. NOTE this glitch
still occurs every time the file is opened. Need to find root.