STL fix
Resave all STLs via applying all booleans in a separate, throw away
file. This avoids what appears to be a glitch in Blender where modifiers
aren't applied correctly on opening a file. This will be investigated
further. Doing this highlighted a few small errors in creases in the
fingers which were corrected too.