Adjust stiffness by finger mass
The 4th finger usually prints well so it was taken as a base. The other
joints were stiffened proportionally to their fingers mass. For example,
the 2nd finger is 1.6 x more massive and therefore the 2nd finger joint
was made 1.6 x stiffer. This was done by widening the two branches of
the joint in the fingers local Y axis, moving the inner faces inwards.
This approach was taken as joint stiffness is less sensitive to changes
in branch width (compared to height) and so changing width can be used
as a "fine tuner".